move template file
